copied entries from old text-file to new Text-file (T00102-107)
I still have to add the Acts for those text-entries.
added persons P0176-179 and will add the bio. data later.
For T00104 I left a gap for Genre see #59
For T00105 the source has to be checked see # 98
I also changed the entry on Yan Wenliang because I had added it twice to the persons- file.