Addition of Social Relations table
I thought about different ways to distinguish interpersonal relations. In the end I got back to the already proposed model of kinship plus another dimension with two possible qualities. This other dimension is friendship in a broad sense.
In the end this category should encompass all relations except for kinship relations. We have city friends and zhiqing comrades, but not seldom they might have been the same guys or became city friends after they returned home. In our texts we once in a while might also encounter friends in the army or party, but this in the end is about the same. The decisive quality at that time was trust, so we could also use the term confidant. Another possible way to distinguish this non-kinship relation would be bosom buddy and friend.

see #153