P0050-P0102 correction and addition
Today i worked on person from 50 to 102.
I found some informations for P0099, but I'm not sure if they are true. Here is a link of it :http://blog.people.com.cn/article/1418106378116.html
The same content will be is quoted by many pages, which are also mentioned to this person.