resove conflict, add birth year and birth place
I redid the work after February 16th, and have added  birth (death) year and place.
It's a new branch to resolve the conflicts at the last PR, see #212 , there was a problem with synchronize because of the network connection.

Could you please close the another branch: re-fixing-person-file ? :-) it has problem again with synchronize.